test/perf/requirements.txt
attrs
click
pyparsing
pyyaml
requests
